+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+  * `futhark pkg` now shells out to `curl` for HTTP requests.
+
 ### Removed
 
 ### Changed